backend/requirements.txt
# Backend Dependencies for Phase 2 SaaS

# FastAPI & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Configuration
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
email-validator==2.1.0

# Payments
stripe==7.4.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP Client
httpx==0.25.2

# Environment Variable Management
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Optional: Database Migrations (for production)
alembic==1.12.1

# Optional: Rate limiting
slowapi==0.1.9

# Optional: CORS
python-multipart==0.0.6
